Install obs-studio from ppa
This installs OBS Studio from the OBS ppa. This is a temporary fix
to prevent the CI from failing if a developer uses newer code
then what the unofficial Ubuntu repo provides (27.2.3 on 22.04).
This isn't ideal as it installs the entire program instead of just
dev packages. In the future, we would provide a libobs-dev package
on our ppa.